+++ src/admin/class-admin-pages.php
@@ -86,6 +86,9 @@ if ( ! class_exists( __NAMESPACE__ . '\Admin_Pages' ) ) {
 				},
 				null
 			);
+
+			// Rename the submenu title for the parent page.
+			$GLOBALS['submenu'][ static::PARENT_PAGE_SLUG ][0][0] = 'Settings';
 		}//end add_admin_pages()
 
 		/**